Dockerfiles/srcs/inventory-app/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
psycopg2==2.9.10